Overall project source tree restructure: Removed poller src and header; all is now declared and defined in  drvAsynIseghalService source and header.
